# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.FacilityResponseDTO:
    properties:
      address:
        type: string
      contact_email:
        type: string
      contact_person:
        type: string
      contact_phone:
        type: string
      created_at:
        type: string
      dhis2_code:
        type: string
      district:
        type: string
      emr_system_code:
        type: string
      emr_system_name:
        type: string
      facility_code:
        type: string
      facility_name:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      level_of_care:
        type: string
      pharmacies:
        items:
          $ref: '#/definitions/dto.PharmacyResponseDTO'
        type: array
      region:
        type: string
      updated_at:
        type: string
      zone:
        type: string
    type: object
...
  dto.PatientVisitCreateDTO:
    properties:
      vst_age:
        type: integer
      vst_batch_number:
        type: string
      vst_facility_code:
        type: string
      vst_patient_category:
        type: string
      vst_patient_code:
        type: string
      vst_product_code:
        type: string
      vst_quantity:
        type: number
      vst_regimen_code:
        type: string
      vst_sex:
        type: string
      vst_system_code:
        type: string
      vst_visit_date:
        type: string
    required:
    - vst_age
    - vst_batch_number
    - vst_facility_code
    - vst_patient_category
    - vst_patient_code
    - vst_product_code
    - vst_quantity
    - vst_regimen_code
    - vst_sex
    - vst_system_code
    - vst_visit_date
    type: object
  dto.PatientVisitResponseDTO:
    properties:
      id:
        type: integer
      sync_status:
        type: integer
      validation_status:
        type: integer
      vst_age:
        type: integer
      vst_batch_number:
        type: string
      vst_facility_code:
        type: string
      vst_patient_category:
        type: string
      vst_patient_code:
        type: string
      vst_product_code:
        type: string
      vst_quantity:
        type: number
      vst_regimen_code:
        type: string
      vst_sex:
        type: string
      vst_system_code:
        type: string
      vst_timestamp:
        type: string
      vst_visit_date:
        type: string
    type: object
  dto.PatientVisitUpdateDTO:
    properties:
      vst_quantity:
        type: number
    type: object
  dto.PharmacyCreateDTO:
    properties:
      facility_id:
        type: integer
      is_active:
        type: boolean
      pharmacy_code:
        type: string
      pharmacy_name:
        type: string
      pharmacy_type:
        type: string
    required:
    - facility_id
    - pharmacy_code
    - pharmacy_name
    type: object
  dto.PharmacyResponseDTO:
    properties:
      created_at:
        type: string
      facility:
        $ref: '#/definitions/dto.FacilityResponseDTO'
      facility_id:
        type: integer
      id:
        type: integer
      is_active:
        type: boolean
      pharmacy_code:
        type: string
      pharmacy_name:
        type: string
      pharmacy_type:
        type: string
      updated_at:
        type: string
    type: object
...
  dto.PharmacyUpdateDTO:
    properties:
      is_active:
        type: boolean
      pharmacy_name:
        type: string
      pharmacy_type:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: ict@moh.go.ug
    name: DHI
  description: Backend API for Supply Chain Repository / Bridge
  termsOfService: https://health.go.ug
  title: MoH Supply Chian Repository API
  version: "1.0"
paths:
  /api/v1/patient-visit:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.PatientVisitResponseDTO'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List patient visits
      tags:
      - PatientVisit
  /api/v1/patient-visit/{id}:
    delete:
      parameters:
      - description: Patient Visit ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete patient visit
      tags:
      - PatientVisit
    get:
      parameters:
      - description: Patient Visit ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PatientVisitResponseDTO'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get patient visit by ID
      tags:
      - PatientVisit
    put:
      consumes:
      - application/json
      parameters:
      - description: Patient Visit ID
        in: path
        name: id
        required: true
        type: integer
      - description: Patient visit update payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/dto.PatientVisitUpdateDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PatientVisitResponseDTO'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update patient visit
      tags:
      - PatientVisit
...
  /patient-visit:
    post:
      consumes:
      - application/json
      parameters:
      - description: Patient visits payload
        in: body
        name: payload
        required: true
        schema:
          items:
            $ref: '#/definitions/dto.PatientVisitCreateDTO'
          type: array
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            items:
              $ref: '#/definitions/dto.PatientVisitResponseDTO'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create patient visits
      tags:
      - PatientVisit
  /pharmacies:
    get:
      parameters:
      - description: Filter by facility ID
        in: query
        name: facility_id
        type: integer
      - description: Filter by active status
        in: query
        name: active
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.PharmacyResponseDTO'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List pharmacies
      tags:
      - Pharmacy
    post:
      consumes:
      - application/json
      parameters:
      - description: Pharmacies payload
        in: body
        name: payload
        required: true
        schema:
          items:
            $ref: '#/definitions/dto.PharmacyCreateDTO'
          type: array
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            items:
              $ref: '#/definitions/dto.PharmacyResponseDTO'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create pharmacies
      tags:
      - Pharmacy
  /pharmacies/{id}:
    delete:
      parameters:
      - description: Pharmacy ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete pharmacy
      tags:
      - Pharmacy
    get:
      parameters:
      - description: Pharmacy ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PharmacyResponseDTO'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get pharmacy by ID
      tags:
      - Pharmacy
    put:
      consumes:
      - application/json
      parameters:
      - description: Pharmacy ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/dto.PharmacyUpdateDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PharmacyResponseDTO'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update pharmacy
      tags:
      - Pharmacy
...
swagger: "2.0"